# Conflicts:
#	scripts/Tween/NumTween.cs

+++ scripts/EffectManager.cs
@@ -7,13 +7,19 @@ using System.Net.NetworkInformation;
 public partial class EffectManager{
     public enum Type{
         None,
-        RoundCardBonus,
-		Multiply,
-        RoundRareBonus,
-        Duplicate,
         RoundThreeRow,
-        LastCardBonus
+        RoundRareBonus,
+        LineColor,
+        Mysterious,
+        Multiply,
+        Duplicate,
+        RoundCardBonus,
+        Sound,
+        TimeCardBonus,
+        LastCardBonus,
 	}
+    public List<bool> b_instant_effect=new List<bool>();
+    public List<Texture> effect_imgs=new List<Texture>();
     public List<Tuple<Type,int>> effect_queue=new List<Tuple<Type, int>>();
     string last_card_type_s;
     int last_card_level,last_row,max_last_row;
@@ -24,6 +30,29 @@ public partial class EffectManager{
     // public void reset_card_effect(){
     //     // next_card_bonus=1;
     // }
+    public EffectManager(){
+        List<cfg.Item> card_data=DataLoader.get_card_data();
+        for(int j=0;j<Enum.GetNames(typeof(Type)).Length;j++){
+            bool nfl=false;
+            for(int i=0;i<card_data.Count;i++){
+                if(card_data[i].Effect1==Enum.GetNames(typeof(Type))[j]){
+                    nfl=true;
+                    b_instant_effect.Add(card_data[i].IsInstantEffect);
+                    Debug.WriteLine(card_data[i].IsInstantEffect);
+                    if(!card_data[i].IsInstantEffect&&card_data[i].EffectImg!=""){
+                        effect_imgs.Add(GD.Load<Texture>("res://imgs/buff/"+card_data[i].EffectImg));
+                    }else{
+                        effect_imgs.Add(null);
+                    }
+                }
+            }
+            if(!nfl){
+                b_instant_effect.Add(true);
+                effect_imgs.Add(null);
+            }
+        }
+        
+    }
     public void reset_round_effect(){
         effect_queue.Clear();
         last_card_type_s="";
@@ -37,7 +66,9 @@ public partial class EffectManager{
         for(int i=0;i<effect_queue.Count;i++){
             if(can_apply(n_card_event,cards,effect_queue[i])){
                 apply_single_effect(n_card_event,effect_queue[i]);
-                remove_index.Add(i);
+                if(should_delete(effect_queue[i])){
+                    remove_index.Add(i);
+                }
             }
         }
         for(int i=0;i<remove_index.Count;i++){
@@ -51,6 +82,21 @@ public partial class EffectManager{
             last_card_type_s=n_card_event.card_type_s;
             last_card_level=n_card_event.card_level;
             max_last_row=Math.Max(max_last_row,last_row);
+        }
+    }
+
+    public void apply_instant_effect(){
+        List<int> remove_index=new List<int>();
+        for(int i=0;i<effect_queue.Count;i++){
+            if(is_instant_effect(effect_queue[i])){
+                apply_single_instant_effect(effect_queue[i]);
+                if(should_delete(effect_queue[i])){
+                    remove_index.Add(i);
+                }
+            }
+        }
+        for(int i=0;i<remove_index.Count;i++){
+            effect_queue.RemoveAt(remove_index[i]-i);
         }
     }
 
@@ -73,6 +119,8 @@ public partial class EffectManager{
             }else{
                 return false;
             }
+        }else if(n_effect.Item1==Type.TimeCardBonus){
+            return true;
         }
         return false;
     }
@@ -89,6 +137,9 @@ public partial class EffectManager{
         }else if(n_effect.Item1==Type.LastCardBonus){
             int[] n4=new int[4]{10,15,25,40};
             n_card_event.score+=n4[n_effect.Item2];
+        }else if(n_effect.Item1==Type.TimeCardBonus){
+            int[] n4=new int[4]{3,4,5,6};
+            n_card_event.score+=n4[n_effect.Item2];
         }
     }
 
@@ -101,5 +152,33 @@ public partial class EffectManager{
             }
         }
         return round_score;
+    }
+
+    bool is_instant_effect(Tuple<Type,int> n_effect){
+        return b_instant_effect[(int)n_effect.Item1];
+    }
+
+    void apply_single_instant_effect(Tuple<Type,int> n_effect){
+        Debug.WriteLine((int)n_effect.Item1);
+    }
+
+    bool should_delete(Tuple<Type,int> n_effect){
+        if(n_effect.Item1==Type.TimeCardBonus){
+            return false;
+        }
+        return true;
+    }
+
+    public void remove_a_time_effect(){
+        List<int> remove_index=new List<int>();
+        for(int i=0;i<effect_queue.Count;i++){
+            if(effect_queue[i].Item1==Type.TimeCardBonus){
+                remove_index.Add(i);
+                break;
+            }
+        }
+        for(int i=0;i<remove_index.Count;i++){
+            effect_queue.RemoveAt(remove_index[i]-i);
+        }
     }
 }
